requirements.txt
Django==4.0.2
django-ckeditor==6.2.0
pillow===9.5.0
pylint==2.12.2
pylint-celery==0.3
pylint-django==2.5.1
djangorestframework==3.13.1
requests===2.27.1
mercadopago===2.1.0
pusher===3.3.0